Mem ordering improvement
This change uses fence instructions to ensure that the ordering of the
assembly instructions is not modified in an undesirable way.  The fence
instructions are only in the IR and are discarded by the clang backend.
This change also reuses the pointer in the buffer, which reduces the
instruction count further.