OpenMP support created improper create events for threads and tasks
The OpenMP support routines for creating threads and tasks puts in
create events to imply the relation between the contexts.  However,
the create events all have 0 skew, since they are in the same logical
thread.  But the middle layer looked to the skew value to know whether
it is a creator or created event.