Memory limits are always in force in runtime
The natural memory limit is set to 90% of free memory.  After reaching
this limit, all further buffer allocations are blocked until there are
no queued buffers.  This should permit Contech to exceed available
memory, disk space permitting.  The runtime also provides an approximate
time in which the system was in the blocked state - CT_LIMIT.