Update Contech runtime in preparation for supporting libraries
When linking libraries, most of the contech runtime needs to be included,
but not all of the routines.  For example, the runtime provides a main routine.
Thus by splitting the runtime into two parts, we have ct_runtime that can
link into everything, and ct_main for the main program itself.

Separate issues include maintaining the basic block IDs and other info.