Contech runtime needs to build a .o version for the marker compilation.
In order to foil any inlining attempts, the linker needs to see a .o file and
not a .bc.  By not inlining the calls, the objparser can find the function
calls in the disassembly and identify the associated events.

TODO: Move the marker calls to a separate file and link against it instead.