Contech_wrapper - Add ARM support to wrapper compiler
First, there is a really simple (i.e., hack) check for ARM.  With ARM, the
wrapper compiler with build and instrument LLVM IR.  For linking, it combines
all of the IR files into a single IR, along with the runtime.  Using
opt -always-inline, all of the key instrumentation routines are inlined.
Then this IR is converted to assembly and linked against native libraries.