Contech - Changes to the runtime and trace format for better support of bulk memmory operations
Bulk memory operations are operations like memcpy, memset, etc.  The interface
for creating these operations has been modified to be more efficient and
simpler to process.  Any change to the runtime format also requires a
corresponding change to the event library.  Tangentially, other revisions
to the event library are also committed as relates to debugging and diagnostics,
which require the updates to ct_file.