Runtime - Reorder operations in writing events
In less common events (barrier, create, and join), there is a
call to rdtsc().  This call was previously in the sequence of
stores; however, clang loses the thread buffer state on that
store and has to reload it afterward.  Moving the rdtsc() to
just before the stores removes these extra instructions.