LLVM Pass - Support for basic block ID eliding
These candidate blocks commonly occur at the end of a diamond.  Both
branches unconditionally converge at the final block, so that block's
ID can be elided from the trace.  More commonly, blocks have been
split with unconditionals when there is an atomic, call, or other
operations and therefore the ID cannot be elided.