Runtime - Partial prototype of directly mmaping space in the trace file
This approach was designed to directly allocate space in the trace file and
avoiding the queuing operations with the background thread writer.  Instead,
the kernel will decide when the data is to be written out and manage the
pages accordingly.  However, the performance was significantly worse.

Perhaps caching is to blame?