LLVM Pass - Update Memory Duplication code to account for non constant values
In some cases, there are GEPIs with slightly different values:
a[i] versus a[i - 32].  The i -32 is a separate value that can be
found by walking the use-def chain to find pure values.  Currently,
the code just accounts for simple adds, but could be revised to
address other, more complex operations.