Contech Pass - 64 bit offsets and multivariate loops
Update Contech to generate 64 bit offsets, which is required to
support multivariate loops (and probably future RAC cases).
Multivariate loops support multiple loop invariants, even if
each is dynamic and the values are applied to the base address.
Currently, there is a severe regression in swaptions, from
either this or increasing the search depth of loops.