LLVM Pass - Updated path information for skipping byte
This code puts an extra byte of space so the path ID can be placed
without overlapping the memory operation.  This space must be
handled carefully as much support assumes the size of IDs.  (#199)
Additionally, the final store of the path ID computes a position to
be used by that block, which has to be split.  (#200).